pyproject.toml
[tool.poetry]
name = "mdb_reddit"
version = "0.1.0"
description = ""
authors = ["DomincWrege <dominic.wrege001@stud.fh-dortmund.de>"]

[tool.poetry.dependencies]
python = "^3.6"
praw = "^7.0.0"
kafka-python = "^2.0.1"
pymongo = "^3.10.1"
pprint = "^0.1"
arrow = "^0.15.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.5.2"

[tool.poetry.scripts]
db_writer = "mdb_reddit.db_writer:main"
fetcher = "mdb_reddit.fetcher:main"
testing = "mdb_reddit.reddit_test:main"
connector = "mdb_reddit.monogdb_connector_source:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

